consolidate CSS
