+++ src/App.css
@@ -7,6 +7,47 @@
   align-items: center;
 }
 
+.light-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+  height: 400px;
+  width: 100px;
+  border: 5px solid black;
+}
+
+.light {
+  border: 3px solid black;
+  border-radius: 50%;
+  width: 75px;
+  height: 75px;
+}
+
+#red-light-off {
+  background-color: rgba(255, 0, 0, 0.403);
+}
+
+#red-light-on {
+  background-color: rgba(255, 0, 0, 1);
+}
+
+#yellow-light-off {
+  background-color: rgba(255, 255, 0, 0.254);
+}
+
+#yellow-light-on {
+  background-color: rgba(255, 255, 0, 1);
+}
+
+#green-light-off {
+  background-color: rgba(0, 128, 0, 0.368);
+}
+
+#green-light-on {
+  background-color: rgba(0, 128, 0, 1);
+}
+
 .start-button {
   margin-top: 10px;
   height: 50px;
