+++ src/components/Button.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import '../App.css'
 
 function Button({ isRunning, onClick }) {
   return <button onClick={onClick}>{isRunning ? 'Stop' : 'Start'}</button>
